[google_adsense] Tighten exports and docs.
