+++ packages/google_adsense/lib/experimental/adsense.dart
@@ -2,8 +2,4 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'src/core/core.dart';
-export 'src/core/core.dart';
-
-/// The singleton instance of the AdSense SDK.
-final AdSense adSense = AdSense();
+export '../src/adsense/adsense.dart';
