--- packages/google_adsense/lib/experimental/google_adsense.dart
@@ -1,7 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-export '../src/ad_unit_configuration.dart';
-export '../src/ad_unit_params.dart';
-export '../src/adsense_stub.dart'
-    if (dart.library.js_interop) '../src/adsense_web.dart';
