+++ packages/google_adsense/lib/src/adsense/adsense.dart
@@ -1,0 +1,8 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'ad_unit_configuration.dart';
+export 'ad_unit_params.dart' hide AdStatus, AdUnitParams;
+export 'adsense_extension_stub.dart'
+    if (dart.library.js_interop) 'adsense_extension.dart';
