+++ packages/google_adsense/lib/src/adsense/adsense_js_interop.dart
@@ -1,0 +1,15 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:js_interop';
+
+import '../js_interop/adsbygoogle.dart';
+
+/// Adds a `requestAd` method to request an AdSense ad.
+extension AdsByGoogleExtension on AdsByGoogle {
+  /// Convenience method for invoking push() with an empty object
+  void requestAd() {
+    push(JSObject());
+  }
+}
