+++ packages/google_adsense/lib/src/core/adsense_web.dart
@@ -7,30 +7,25 @@ import 'dart:js_interop';
 import 'package:flutter/widgets.dart';
 import 'package:web/web.dart' as web;
 
-import 'ad_unit_configuration.dart';
-import 'ad_unit_widget.dart';
-import 'js_interop/adsbygoogle.dart' show adsbygooglePresent;
-import 'js_interop/package_web_tweaks.dart';
+import '../js_interop/adsbygoogle.dart' show adsbygooglePresent;
+import '../js_interop/package_web_tweaks.dart';
 
-import 'logging.dart';
+import '../logging.dart';
 
-/// Returns a singleton instance of Adsense library public interface
-final AdSense adSense = AdSense();
-
-/// Main class to work with the library
+/// The web implementation of the AdSense API.
 class AdSense {
   bool _isInitialized = false;
 
-  /// The ad client ID used by this client.
-  late String _adClient;
+  /// The [Publisher ID](https://support.google.com/adsense/answer/2923881).
+  late String adClient;
   static const String _url =
       'https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-';
 
   /// Initializes the AdSense SDK with your [adClient].
   ///
-  /// Should be called ASAP, ideally in the main method of your app.
+  /// The [adClient] parameter is your AdSense [Publisher ID](https://support.google.com/adsense/answer/2923881).
   ///
-  /// Noops after the first call.
+  /// Should be called ASAP, ideally in the `main` method.
   void initialize(
     String adClient, {
     @visibleForTesting bool skipJsLoader = false,
@@ -40,24 +35,19 @@ class AdSense {
       debugLog('adSense.initialize called multiple times. Skipping init.');
       return;
     }
-    _adClient = adClient;
+    this.adClient = adClient;
     if (!(skipJsLoader || _sdkAlreadyLoaded(testingTarget: jsLoaderTarget))) {
-      _loadJsSdk(_adClient, jsLoaderTarget);
+      _loadJsSdk(adClient, jsLoaderTarget);
     } else {
       debugLog('SDK already on page. Skipping init.');
     }
     _isInitialized = true;
   }
 
-  /// Returns an [AdUnitWidget] with the specified [configuration].
-  Widget adUnit(AdUnitConfiguration configuration) {
-    return AdUnitWidget(adClient: _adClient, configuration: configuration);
-  }
-
   bool _sdkAlreadyLoaded({
     web.HTMLElement? testingTarget,
   }) {
-    final String selector = 'script[src*=ca-pub-$_adClient]';
+    final String selector = 'script[src*=ca-pub-$adClient]';
     return adsbygooglePresent ||
         web.document.querySelector(selector) != null ||
         testingTarget?.querySelector(selector) != null;
